fix: ensure binary with try-runtime feature
