+++ crates/pop-cli/src/commands/test/on_runtime_upgrade.rs
@@ -8,7 +8,7 @@ use crate::{
 	common::{
 		builds::guide_user_to_select_profile,
 		prompt::display_message,
-		runtime::ensure_runtime_binary_exists,
+		runtime::{ensure_runtime_binary_exists, Feature},
 		try_runtime::{
 			argument_exists, check_try_runtime_and_prompt, collect_shared_arguments,
 			partition_arguments, ArgumentConstructor,
@@ -139,6 +139,7 @@ impl TestOnRuntimeUpgradeCommand {
 				cli,
 				&current_dir().unwrap_or(PathBuf::from("./")),
 				self.profile.as_ref().ok_or_else(|| anyhow::anyhow!("No profile provided"))?,
+				vec![Feature::TryRuntime],
 				!self.no_build,
 			)?);
 		}
